student-submission/requirements.txt
creevo-agent-library[mcp]
python-dotenv
